Combine collection and token under the digitalAsset namespace
