+++ src/api/digitalAsset.ts
@@ -1,21 +1,36 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
+import {
+  GetCollectionDataResponse,
+  GetCurrentTokenOwnershipResponse,
+  GetOwnedTokensResponse,
+  GetTokenActivityResponse,
+  GetTokenDataResponse,
+  HexInput,
+  OrderBy,
+  PaginationArgs,
+  TokenStandard,
+} from "../types";
 import { AptosConfig } from "./aptosConfig";
+import { Account } from "../core";
+import { GenerateTransactionOptions, SingleSignerTransaction } from "../transactions/types";
 import {
   CreateCollectionOptions,
   createCollectionTransaction,
-  getCollectionId,
   getCollectionData,
-} from "../internal/collection";
-import { Account } from "../core";
-import { GenerateTransactionOptions, SingleSignerTransaction } from "../transactions/types";
-import { GetCollectionDataResponse, HexInput, TokenStandard } from "../types";
+  getCollectionId,
+  getCurrentTokenOwnership,
+  getOwnedTokens,
+  getTokenActivity,
+  getTokenData,
+  mintTokenTransaction,
+} from "../internal/digitalAsset";
 
 /**
- * A class to query all `Collection` related queries on Aptos.
+ * A class to query all `DigitalAsset` related queries on Aptos.
  */
-export class Collection {
+export class DigitalAsset {
   readonly config: AptosConfig;
 
   constructor(config: AptosConfig) {
@@ -99,5 +114,79 @@ export class Collection {
     };
   }): Promise<string> {
     return getCollectionId({ aptosConfig: this.config, ...args });
+  }
+
+  /**
+   * Create a transaction to mint a token into the creators account within an existing collection.
+   *
+   * @param args.creator the creator of the collection
+   * @param args.collection the name of the collection the token belongs to
+   * @param args.description the description of the token
+   * @param args.name the name of the token
+   * @param args.uri the URI to additional info about the token
+   *
+   * @returns A SingleSignerTransaction that can be simulated or submitted to chain
+   */
+  async mintTokenTransaction(args: {
+    creator: Account;
+    collection: string;
+    description: string;
+    name: string;
+    uri: string;
+    options?: GenerateTransactionOptions;
+  }): Promise<SingleSignerTransaction> {
+    return mintTokenTransaction({ aptosConfig: this.config, ...args });
+  }
+
+  /**
+   * Gets token data given the address of a token.
+   *
+   * @param args.tokenAddress The address of the token
+   * @returns GetTokenDataResponse containing relevant data to the token.
+   */
+  async getTokenData(args: { tokenAddress: HexInput }): Promise<GetTokenDataResponse> {
+    return getTokenData({ aptosConfig: this.config, ...args });
+  }
+
+  /**
+   * Gets token ownership data given the address of a token.
+   *
+   * @param args.tokenAddress The address of the token
+   * @returns GetCurrentTokenOwnershipResponse containing relevant ownership data of the token.
+   */
+  async getCurrentTokenOwnership(args: { tokenAddress: HexInput }): Promise<GetCurrentTokenOwnershipResponse> {
+    return getCurrentTokenOwnership({ aptosConfig: this.config, ...args });
+  }
+
+  /**
+   * Gets the tokens that the given address owns.
+   *
+   * @param args.ownerAddress The address of the owner
+   * @returns GetOwnedTokensResponse containing ownership data of the tokens belonging to the ownerAddresss.
+   */
+  async getOwnedTokens(args: {
+    ownerAddress: HexInput;
+    options?: {
+      pagination?: PaginationArgs;
+      orderBy?: OrderBy<GetOwnedTokensResponse[0]>;
+    };
+  }): Promise<GetOwnedTokensResponse> {
+    return getOwnedTokens({ aptosConfig: this.config, ...args });
+  }
+
+  /**
+   * Gets the activity data given the address of a token.
+   *
+   * @param args.tokenAddress The address of the token
+   * @returns GetTokenActivityResponse containing relevant activity data to the token.
+   */
+  async getTokenActivity(args: {
+    tokenAddress: HexInput;
+    options?: {
+      pagination?: PaginationArgs;
+      orderBy?: OrderBy<GetTokenActivityResponse[0]>;
+    };
+  }): Promise<GetTokenActivityResponse> {
+    return getTokenActivity({ aptosConfig: this.config, ...args });
   }
 }
